refactor(loon-core): split storage checkpoint and wal modules
