+++ crates/loon-core/src/storage/mod.rs
@@ -1,0 +1,1 @@
+pub(crate) mod content;
